Fix comment author IP storage overflow
IPv4 addresses are 32 bits, but we need the full range of those 32 bits,
which we don't get with signed ints. Rather than pack it in there with 0
starting at -2**32/2, I'm just switching to a BIGINT field type.

This will need to be changed later to support 128 bit IPv6 addresses.

+++ mediacore/model/comments.py
@@ -24,7 +24,7 @@ Comments come with two status flags:
 """
 from datetime import datetime
 from sqlalchemy import Table, ForeignKey, Column, sql
-from sqlalchemy.types import Unicode, UnicodeText, Integer, DateTime, Boolean, Float
+from sqlalchemy.types import BigInteger, Boolean, DateTime, Integer, Unicode, UnicodeText
 from sqlalchemy.orm import mapper, relation, backref, synonym, composite, column_property, validates, interfaces, Query
 
 from mediacore.model import AuthorWithIP
@@ -42,7 +42,7 @@ comments = Table('comments', metadata,
     Column('publishable', Boolean, default=False, nullable=False),
     Column('author_name', Unicode(50), nullable=False),
     Column('author_email', Unicode(255)),
-    Column('author_ip', Integer, nullable=False),
+    Column('author_ip', BigInteger, nullable=False),
     Column('body', UnicodeText, nullable=False),
     mysql_engine='InnoDB',
     mysql_charset='utf8',
